requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
pinecone-client>=5.0.0
pypdf>=4.2.0
python-dotenv>=1.0.1
streamlit>=1.36.0
unstructured>=0.15.0
unstructured[pdf]>=0.15.0
docx2txt>=0.8
